refactor: finalized directory structure and naming conventions
