--- src/core/types/actionObject.d.ts
@@ -1,3 +1,0 @@
-export type ActionObject = {
-	actionType: number
-};
