--- src/core/types/messageObject.d.ts
@@ -1,4 +1,0 @@
-export type MessageObject = {
-	messageType: string,
-	message: string
-};
