--- src/core/types/messageObjectArray.d.ts
@@ -1,3 +1,0 @@
-import { MessageObject } from "./messageObject";
-
-export type MessageObjectArray = MessageObject[];
