+++ src/core/types/shared.d.ts
@@ -1,0 +1,10 @@
+export type MessageObject = {
+	messageType: string,
+	message: string
+};
+
+export type MessageObjectArray = MessageObject[];
+
+export type ActionObject = {
+	actionType: number
+};
