+++ src/enums/pluginBranch.ts
@@ -1,4 +1,4 @@
-enum branchType {
+export enum branchType {
 	development = "Development",
 	nightly = "Nightly",
 	production = "Production"
