+++ tsconfig.json
@@ -2,10 +2,10 @@
 	"compilerOptions": {
 		"module": "commonjs",
 		"target": "ES2020",
-		"outDir": "out",
+		"outDir": "dist",
 		"lib": [ "ES2020" ],
 		"sourceMap": true,
-		"rootDir": "source",
+		"rootDir": "src",
 		"strict": true   /* enable all strict type-checking options */
 		/* Additional Checks */
 		// "noImplicitReturns": true, /* Report error when not all code paths in function return a value. */
